chore(formatter): move JsAssignmentWithDefault to its own file
