+++ crates/rome_formatter/src/ts/assignment/mod.rs
@@ -1,6 +1,7 @@
 mod any_assignment;
 mod any_assignment_pattern;
 mod array_assignment_pattern;
+mod assignment_with_default;
 mod computed_member_assignment;
 mod identifier_assignment_target;
 mod object_assignment_pattern;
